Reduce mouse steering net spam
Reduces the amount of net message spam to the server by only sending when the player is actually in a simfphys vehicle and is actively using mouse steering.

This could probably stand to be improved further because it still sends a lot of messages when the conditions are met, but the overall impact should still be drastically lower.